Some minor modifications for style.  Introduction of a GridPos class to capture the x,y coordinate system in a compact way. The name is specific due to the current use but we can always make it more generic.
Added a cell manager which will be responsible for converting
canonical integer representations of the drawn cells into the
images.  Placeholder present in 'ACellManagerSimple' but
this will obviously need to mirror the AMapReader in that we
will read cells from a manifest file and then the images from
either sprite sheets or simple images on disk.