Adding integration test capabilities via Mockito into our code base.
Today Mockito appears to be the current winner in terms of best mock framework.  A good analysis of this, versus easymock and jmockit is here:
    https://www.baeldung.com/mockito-vs-easymock-vs-jmockit

I also added a simple test for the MapInstance which
begins to show how interaction based testing works.

This is just a starting point but a good commit unit
as getting it all wired up and an initial test took
a bit of time!  :)