Separated out the concept of a player character from the map view and the generic concept as a person playing the game with the intrinsic properties that will exist everywhere.
Made the resource manager be a generic and got rid of some boilerplate
code and some classes - smaller code - yay!

Created a Cell Manager which is driven by a sprite sheet.  The sheet
is loaded up on startup and the cells are automatically extracted
from the sheet based on parameters.